feat(21068): remove action queue for new MV3
