--- test/jest/background.js
@@ -1,5 +1,0 @@
-import { _setBackgroundConnection } from '../../ui/store/action-queue';
-
-export const setBackgroundConnection = (backgroundConnection = {}) => {
-  _setBackgroundConnection(backgroundConnection);
-};
