+++ test/jest/index.js
@@ -1,6 +1,5 @@
 export { screen, fireEvent, waitFor } from '@testing-library/react';
 export { createSwapsMockStore } from './mock-store';
 export { renderWithProvider } from './rendering';
-export { setBackgroundConnection } from './background';
 export * as MOCKS from './mocks';
 export * as CONSTANTS from './constants';
