+++ ui/pages/confirm-send-ether/confirm-send-ether.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { renderWithProvider } from '../../../test/lib/render-helpers';
-import { setBackgroundConnection } from '../../../test/jest';
+import { setBackgroundConnection } from '../../store/background-connection';
 import mockState from '../../../test/data/mock-state.json';
 import configureStore from '../../store/store';
 import ConfirmSendEther from './confirm-send-ether';
