+++ ui/pages/confirm-transaction/confirm-transaction.transaction.test.js
@@ -3,7 +3,7 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 
 import { renderWithProvider } from '../../../test/lib/render-helpers';
-import { setBackgroundConnection } from '../../../test/jest';
+import { setBackgroundConnection } from '../../store/background-connection';
 import mockState from '../../../test/data/mock-state.json';
 import {
   CONFIRM_SEND_ETHER_PATH,
