+++ ui/pages/onboarding-flow/pin-extension/pin-extension.test.js
@@ -3,10 +3,8 @@ import { fireEvent } from '@testing-library/react';
 import reactRouterDom from 'react-router-dom';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import {
-  renderWithProvider,
-  setBackgroundConnection,
-} from '../../../../test/jest';
+import { setBackgroundConnection } from '../../../store/background-connection';
+import { renderWithProvider } from '../../../../test/jest';
 import PinExtension from './pin-extension';
 
 const completeOnboardingStub = jest
