+++ ui/pages/send/send.test.js
@@ -5,15 +5,13 @@ import { useLocation } from 'react-router-dom';
 import { NetworkType } from '@metamask/controller-utils';
 import { SEND_STAGES, startNewDraftTransaction } from '../../ducks/send';
 import { domainInitialState } from '../../ducks/domains';
+import { setBackgroundConnection } from '../../store/background-connection';
 import {
   CHAIN_IDS,
   GOERLI_DISPLAY_NAME,
   NETWORK_TYPES,
 } from '../../../shared/constants/network';
-import {
-  renderWithProvider,
-  setBackgroundConnection,
-} from '../../../test/jest';
+import { renderWithProvider } from '../../../test/jest';
 import { GasEstimateTypes } from '../../../shared/constants/gas';
 import { KeyringType } from '../../../shared/constants/keyring';
 import { INITIAL_SEND_STATE_FOR_EXISTING_DRAFT } from '../../../test/jest/mocks';
