+++ ui/pages/swaps/create-new-swap/create-new-swap.test.js
@@ -2,11 +2,11 @@ import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 
+import { setBackgroundConnection } from '../../../store/background-connection';
 import {
   renderWithProvider,
   createSwapsMockStore,
   fireEvent,
-  setBackgroundConnection,
 } from '../../../../test/jest';
 import {
   setSwapsFromToken,
